src/test/java/edu/ithaca/dragon/practice/FunctionPracticeTest.java
package edu.ithaca.dragon.practice;

import static org.junit.jupiter.api.Assertions.assertEquals;
//import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertThrows;
//import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.junit.jupiter.api.Assertions.fail;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import org.junit.jupiter.api.Test;

public class FunctionPracticeTest {

    @Test
    public void largestOfThreeTest(){
        assertEquals(5, FunctionPractice.largestOfThree(4, 5, 2));
        assertEquals(5, FunctionPractice.largestOfThree(5, 4, 2));
        assertEquals(5, FunctionPractice.largestOfThree(4, 2, 5));
        assertEquals(5, FunctionPractice.largestOfThree(5, 5, 5)); //these test switch the numbers around to make sure the result is correct reguardless of where the biggest number is 

        assertThrows(IllegalArgumentException.class, () -> FunctionPractice.largestOfThree(-1, -1, -1));
        assertThrows(IllegalArgumentException.class, () -> FunctionPractice.largestOfThree(5, -1, 3));
        assertThrows(IllegalArgumentException.class, () -> FunctionPractice.largestOfThree(5, 3, -1));
    }

    @Test
    public void calcSalePriceTest(){ 
        assertEquals(11, FunctionPractice.calcSalePrice(20, 50, 10));
       //fail("Not implemented yet");
    }

    @Test
    public void isGoodDogTest(){
        assertEquals(true, FunctionPractice.isGoodDog(1, 1, true));
        assertEquals(false, FunctionPractice.isGoodDog(1, 0, true));

    }

    @Test
    public void findFirstLargest(){ 
        List<Integer> numbers2 = Arrays.asList(0,1,2,3);
        assertEquals(3, FunctionPractice.findFirstLargest(numbers2));
        
    }
    
    
}
